Removed comments amount, fixed history room card sizing
