+++ app/src/main/java/com/example/roomer/presentation/screens/navbar_screens/home_screen/HomeScreen.kt
@@ -206,7 +206,6 @@ private fun RecentlyWatched(
         LazyRow(
             modifier = Modifier
                 .fillMaxWidth()
-                .height(148.dp)
                 .padding(top = 8.dp),
             horizontalArrangement = Arrangement.spacedBy(
                 dimensionResource(id = R.dimen.list_elements_margin)
